# Speed up emit_dist with precomputed length codings for static trees