Benchmark results:
```
Benchmark 1 (60 runs): ./compress-baseline 1 rs silesia-small.tar
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          83.2ms ± 2.00ms    79.7ms … 93.9ms          1 ( 2%)        0%
  peak_rss           26.7MB ± 63.7KB    26.6MB … 26.7MB          0 ( 0%)        0%
  cpu_cycles          292M  ±  648K      291M  …  294M           0 ( 0%)        0%
  instructions        612M  ±  324       612M  …  612M           0 ( 0%)        0%
  cache_references    401K  ± 8.39K      396K  …  463K           6 (10%)        0%
  cache_misses        302K  ± 5.51K      277K  …  312K           8 (13%)        0%
  branch_misses      3.05M  ± 4.12K     3.04M  … 3.06M           0 ( 0%)        0%
Benchmark 2 (62 runs): ./target/release/examples/compress 1 rs silesia-small.tar
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          81.1ms ± 1.14ms    78.7ms … 84.4ms          1 ( 2%)        ⚡-  2.5% ±  0.7%
  peak_rss           26.7MB ± 77.6KB    26.5MB … 26.7MB          0 ( 0%)          -  0.0% ±  0.1%
  cpu_cycles          285M  ±  641K      283M  …  287M           1 ( 2%)        ⚡-  2.6% ±  0.1%
  instructions        590M  ±  286       590M  …  590M           0 ( 0%)        ⚡-  3.7% ±  0.0%
  cache_references    401K  ± 3.52K      397K  …  420K           3 ( 5%)          -  0.2% ±  0.6%
  cache_misses        302K  ± 7.85K      255K  …  316K           8 (13%)          +  0.0% ±  0.8%
  branch_misses      2.87M  ± 3.09K     2.86M  … 2.87M           1 ( 2%)        ⚡-  5.9% ±  0.0%
```

Notes:
- The improvement is smaller for higher compression levels.
- Possibly controversial: I changed a couple of int-packing operations from big-endian to little-endian, which further improved the performance on x86. Many, but not all, computers are little-endian these days, but I don't know if this will cause a regression on systems that aren't.