@folkertdev 2024-12-14
it's late here so a proper review will have to wait till tomorrow, but this is looking really good!

```
Benchmark 1 (64 runs): target/release/examples/compress-baseline 1 rs silesia-small.tar
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          78.2ms ± 1.41ms    76.6ms … 82.2ms          0 ( 0%)        0%
  peak_rss           26.7MB ± 65.2KB    26.6MB … 26.7MB          0 ( 0%)        0%
  cpu_cycles          286M  ± 5.95M      280M  …  304M           2 ( 3%)        0%
  instructions        590M  ±  419       590M  …  590M           2 ( 3%)        0%
  cache_references   20.0M  ±  265K     19.8M  … 21.8M           1 ( 2%)        0%
  cache_misses        523K  ± 68.4K      423K  …  726K           1 ( 2%)        0%
  branch_misses      3.05M  ± 8.30K     3.04M  … 3.10M           1 ( 2%)        0%
Benchmark 2 (66 runs): target/release/examples/blogpost-compress 1 rs silesia-small.tar
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          76.7ms ± 1.55ms    75.8ms … 88.6ms          4 ( 6%)        ⚡-  1.9% ±  0.7%
  peak_rss           26.7MB ± 63.5KB    26.6MB … 26.7MB          0 ( 0%)          +  0.0% ±  0.1%
  cpu_cycles          280M  ± 7.14M      275M  …  335M           4 ( 6%)        ⚡-  2.4% ±  0.8%
  instructions        567M  ±  264       567M  …  567M           2 ( 3%)        ⚡-  3.8% ±  0.0%
  cache_references   20.0M  ±  515K     19.7M  … 23.8M           6 ( 9%)          -  0.2% ±  0.7%
  cache_misses        459K  ± 96.8K      343K  …  904K           3 ( 5%)        ⚡- 12.2% ±  5.5%
  branch_misses      2.93M  ± 5.41K     2.92M  … 2.94M           0 ( 0%)        ⚡-  4.1% ±  0.1%
```

> The improvement is smaller for higher compression levels.

That makes sense; they do more other work, so the emit functions are less of a bottleneck.

> I changed a couple of int-packing operations from big-endian to little-endian

From a cursory look, you only use those operations on data from the tables right? That's fine: changing the endianness  would be problematic when reading from the input stream, but if you just store the data in the opposite order then that's all good. We also test on a big-endian target (s390x) on CI, so it would catch any regressions.