@brian-pane 2024-12-15
Right, I only changed `from_be_bytes` to `from_le_bytes` in functions that were creating integers from temporary arrays constructed specifically for that purpose. Ideally, we could choose at compile time between LE and BE implementations to match the target architecture, but I don't know how to do that in Rust.