@folkertdev 2024-12-15
> Ideally, we could choose at compile time between LE and BE implementations to match the target architecture, but I don't know how to do that in Rust.

there are helpers for using the native endianness

- https://doc.rust-lang.org/std/primitive.u16.html#method.from_ne_bytes
- https://doc.rust-lang.org/std/primitive.u16.html#method.to_ne_bytes

So as long as you make sure uses of these functions are "in-sync" you should be good.