@folkertdev 2024-12-15
i added `encode_dist` to mirror `encode_len`. I don't think we can (profitable) create a cached lookup table here, but at least this reduces duplication and reduces data dependencies. On my machine there is no change in performance with/without this refactor though, so it's mostly the deduplication that matters there.

Btw, you can `git pull --rebase` to fetch changes that maintainers make onto your own local branch. That prevents a merge commit from being created (which we'd like to avoid).

I think this looks good now, we'll likely merge this tomorrow, but first I need to make a release (our releases get audited, and someone did that before the weekend, so it's simpler to get this PR added to the next release).